Added new file for search functions. Restructured update.py and update_functions.py
