+++ src/functions/update_functions.py
@@ -1,11 +1,10 @@
 #===============================================================================
 #                               Update Functions
 #===============================================================================
-import csv
-import re
 from colorama import Style
 from . import csv_functions
-from .. import global_vars, model
+from .. import global_vars
+from ..functions import search_functions
 
 job_categories = global_vars.job_categories
 status_options = global_vars.status_options
@@ -20,58 +19,6 @@ section_prompt = """
     3: NOTES
 -----------------------------
             """
-
-### Search spreadsheet for existing job listings. Create master list of all listings
-def find_job(args):
-    master = []
-    matches = []
-    with open(global_vars.f_name,"r") as spreadsheet:
-        csv_file = csv.reader(spreadsheet,delimiter=",")
-        line = 0
-        next(csv_file)
-        for row in csv_file:
-            if re.search("%s" % str(args.update[0]),"%s" % row[1],re.I):
-                match = [model.Job(row[0],row[1],row[2],row[3],row[4]),line]
-                matches.append(match)
-            job = dict(zip(job_categories,row))
-            master.append(job)
-            line += 1
-
-    return master,matches
-
-### Set formatting for header
-def set_print_format(matches):
-    c_len = 0
-    t_len = 0
-    n_len = 0
-    for match in matches:
-        if len(match[0].company) > c_len:
-            c_len = len(match[0].company) + 2
-        if len(match[0].title) > t_len:
-            t_len = len(match[0].title) + 2 if len(match[0].title) > 10 else len(job_categories[2]) + 2
-        if len(match[0].notes) > n_len:
-            n_len = len(match[0].notes) + 2
-
-    return c_len,t_len,n_len
-
-### Print all matches
-def print_matches(matches):
-    c_len,t_len,n_len = set_print_format(matches)
-    found_header = f"\n{'='*61} EXISTING JOB LISTINGS {'='*62}"
-    table_header = f"\nNumber {job_categories[0]:<{19}} {job_categories[1]:<{c_len}} {job_categories[2]:<{t_len}} {job_categories[3]:<{16}} {job_categories[4]:<{n_len}}"
-    print(found_header)
-    print(table_header)
-    print("-"*len(table_header))
-    
-    n = 0
-    for match in matches:
-        details = [match[0].date,match[0].company,match[0].title,match[0].status,match[0].notes]
-        description = f"{n:<6} {details[0]:<{19}} {details[1]:<{c_len}} {details[2]:<{t_len}} {details[3]:<{16}} {details[4]:<{n_len}}\n"
-        global_vars.set_color(description,details,c_len,t_len,n_len)
-
-        n += 1
-
-    return n
 
 ### Select a match
 def select_job(matches,n):
@@ -129,7 +76,7 @@ def update_section(section,matches,selected):
 
 ### Header for listing changes
 def list_changes_header(matches,selected):
-    c_len,t_len,n_len = set_print_format(matches)
+    c_len,t_len,n_len = search_functions.set_print_format(matches)
     details = [matches[selected][0].date,matches[selected][0].company,matches[selected][0].title,matches[selected][0].status,matches[selected][0].notes]
     description = f"{details[0]:<{19}} {details[1]:<{c_len}} {details[2]:<{t_len}} {details[3]:<{16}} {details[4]:<{n_len}}\n"
     table_header = f"\n{job_categories[0]:<{19}} {job_categories[1]:<{c_len}} {job_categories[2]:<{t_len}} {job_categories[3]:<{16}} {job_categories[4]:<{n_len}}"
